+++ src/main/java/com/example/fms/service/CounterpartyServiceImpl.java
@@ -1,14 +1,13 @@
 package com.example.fms.service;
 
 import com.example.fms.dto.CounterpartyDTO;
-import com.example.fms.entity.Counterparty;
-import com.example.fms.entity.Journal;
-import com.example.fms.entity.ResponseMessage;
-import com.example.fms.entity.Transaction;
+import com.example.fms.entity.*;
 import com.example.fms.exception.ResourceNotFoundException;
 import com.example.fms.repository.CounterpartyRepository;
 import com.example.fms.repository.JournalRepository;
 import com.example.fms.repository.UserRepository;
+import org.hibernate.Filter;
+import org.hibernate.Session;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -17,6 +16,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityManager;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
@@ -29,10 +29,17 @@ public class CounterpartyServiceImpl implements CounterpartyService{
     private JournalRepository journalRepository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private EntityManager entityManager;
 
     @Override
-    public List<Counterparty> getAll() {
-        return counterpartyRepository.findAll();
+    public List<Counterparty> getAll(boolean isDeleted) {
+        Session session = entityManager.unwrap(Session.class);
+        Filter filter = session.enableFilter("deletedCounterpartyFilter");
+        filter.setParameter("isDeleted", isDeleted);
+        List<Counterparty> counterparties = counterpartyRepository.findAll();
+        session.disableFilter("deletedCounterpartyFilter");
+        return counterparties;
     }
 
     @Override
@@ -78,6 +85,8 @@ public class CounterpartyServiceImpl implements CounterpartyService{
     public ResponseEntity<Counterparty> getCounterpartyById(Long id) {
         Counterparty counterparty = counterpartyRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Counterparty id " + id + " not found!"));
+        if (counterparty.isDeleted())
+            throw new ResourceNotFoundException("Counterparty id " + id + " was deleted!");
         return ResponseEntity.ok().body(counterparty);
     }
 
@@ -85,6 +94,8 @@ public class CounterpartyServiceImpl implements CounterpartyService{
     public ResponseEntity<Counterparty> updateCounterpartyById(CounterpartyDTO counterpartyDTO, Long id, String userEmail) {
          Counterparty result = counterpartyRepository.findById(id)
                 .map(counterparty -> {
+                    if (counterparty.isDeleted())
+                        throw new ResourceNotFoundException("Counterparty id " + id + " was deleted!");
                     counterparty.setName(counterpartyDTO.getName());
                     return counterpartyRepository.save(counterparty);
                 })
@@ -105,7 +116,10 @@ public class CounterpartyServiceImpl implements CounterpartyService{
         Counterparty counterparty = counterpartyRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Counterparty id " + id + " not found!"));
 
-            counterpartyRepository.deleteById(id);
+        if (counterparty.isDeleted())
+            throw new ResourceNotFoundException("Counterparty id " + id + " was deleted!");
+
+        counterpartyRepository.deleteById(id);
             Journal journal = new Journal();
             journal.setTable("COUNTERPARTY: " + counterparty.getName());
             journal.setAction("delete");
